+++ server/api/controllers/toolsController.js
@@ -1,5 +1,5 @@
 import mongoose from "mongoose";
-import models from "../models/ToolsModel.js";
+import models from "../models/product.js";
 const {Tools, Gen} = models
 
 export const procureTools = async (req, res, next) => {
